fix tests and only clean up in a non-interactive environment
also add the entry to the NEWS

+++ tests/testthat/setup.R
@@ -1,6 +1,6 @@
 cleanup <- function() {
   cachedir <- tools::R_user_dir("snvecR", which = "cache")
-  if (dir.exists(cachedir)) {
+  if (dir.exists(cachedir) && !interactive()) {
     cli::cli_inform("Removing {.file {cachedir}} from reproducible environment.")
     unlink(cachedir, recursive = TRUE)
   }
